i18n をモジュール化
