+++ i18n_support/__init__.py
@@ -1,0 +1,5 @@
+"""Internal modules for i18n behavior.
+
+The public import surface remains i18n.py. Keep application code importing from
+that compatibility module unless there is a reason to depend on internals.
+"""
