+++ i18n_support/constants.py
@@ -1,0 +1,253 @@
+from contextvars import ContextVar
+
+SUPPORTED_LANGUAGES = (
+    "ja",
+    "en",
+    "zh-CN",
+    "zh-TW",
+    "ko",
+    "fr",
+    "es",
+    "de",
+    "pt",
+    "it",
+    "vi",
+    "th",
+    "id",
+    "tr",
+    "uk",
+    "ru",
+    "nl",
+    "hi",
+    "bn",
+    "pl",
+    "sw",
+    "ar",
+)
+DEFAULT_LANGUAGE = "ja"
+LANGUAGE_COOKIE_NAME = "fsqr_language"
+LANGUAGE_COOKIE_MAX_AGE_SECONDS = 365 * 24 * 60 * 60
+current_language_ctx: ContextVar[str] = ContextVar("current_language", default="ja")
+
+LANGUAGE_OPTIONS = (
+    {"code": "ja", "label": "日本語", "flag": "🇯🇵"},
+    {"code": "en", "label": "English", "flag": "🇺🇸"},
+    {"code": "zh-CN", "label": "简体中文", "flag": "🇨🇳"},
+    {"code": "zh-TW", "label": "繁體中文", "flag": "🇹🇼"},
+    {"code": "ko", "label": "한국어", "flag": "🇰🇷"},
+    {"code": "fr", "label": "Français", "flag": "🇫🇷"},
+    {"code": "es", "label": "Español", "flag": "🇪🇸"},
+    {"code": "de", "label": "Deutsch", "flag": "🇩🇪"},
+    {"code": "pt", "label": "Português", "flag": "🇵🇹"},
+    {"code": "it", "label": "Italiano", "flag": "🇮🇹"},
+    {"code": "ru", "label": "Русский", "flag": "🇷🇺"},
+    {"code": "nl", "label": "Nederlands", "flag": "🇳🇱"},
+    {"code": "hi", "label": "हिन्दी", "flag": "🇮🇳"},
+    {"code": "bn", "label": "বাংলা", "flag": "🇧🇩"},
+    {"code": "vi", "label": "Tiếng Việt", "flag": "🇻🇳"},
+    {"code": "th", "label": "ไทย", "flag": "🇹🇭"},
+    {"code": "id", "label": "Bahasa Indonesia", "flag": "🇮🇩"},
+    {"code": "tr", "label": "Türkçe", "flag": "🇹🇷"},
+    {"code": "uk", "label": "Українська", "flag": "🇺🇦"},
+    {"code": "pl", "label": "Polski", "flag": "🇵🇱"},
+    {"code": "sw", "label": "Kiswahili", "flag": "🇹🇿"},
+    {"code": "ar", "label": "العربية", "flag": "🇸🇦"},
+)
+
+COUNTRY_LANGUAGE_MAP = {
+    "JP": "ja",
+    "CN": "zh-CN",
+    "SG": "zh-CN",
+    "HK": "zh-TW",
+    "MO": "zh-TW",
+    "TW": "zh-TW",
+    "KR": "ko",
+    "FR": "fr",
+    "ES": "es",
+    "MX": "es",
+    "AR": "es",
+    "CL": "es",
+    "CO": "es",
+    "PE": "es",
+    "US": "en",
+    "GB": "en",
+    "AU": "en",
+    "CA": "en",
+    "NZ": "en",
+    "DE": "de",
+    "AT": "de",
+    "CH": "de",
+    "PT": "pt",
+    "BR": "pt",
+    "IT": "it",
+    "RU": "ru",
+    "BY": "ru",
+    "KZ": "ru",
+    "KG": "ru",
+    "NL": "nl",
+    "BE": "nl",
+    "IN": "hi",
+    "BD": "bn",
+    "VN": "vi",
+    "TH": "th",
+    "ID": "id",
+    "TR": "tr",
+    "UA": "uk",
+    "PL": "pl",
+    "TZ": "sw",
+    "KE": "sw",
+    "UG": "sw",
+    "SA": "ar",
+    "AE": "ar",
+    "EG": "ar",
+    "JO": "ar",
+    "KW": "ar",
+    "QA": "ar",
+}
+
+HTML_LANG_MAP = {
+    "ja": "ja",
+    "en": "en",
+    "zh-CN": "zh-CN",
+    "zh-TW": "zh-TW",
+    "ko": "ko",
+    "fr": "fr",
+    "es": "es",
+    "de": "de",
+    "pt": "pt",
+    "it": "it",
+    "ru": "ru",
+    "nl": "nl",
+    "hi": "hi",
+    "bn": "bn",
+    "vi": "vi",
+    "th": "th",
+    "id": "id",
+    "tr": "tr",
+    "uk": "uk",
+    "pl": "pl",
+    "sw": "sw",
+    "ar": "ar",
+}
+META_LANGUAGE_MAP = {
+    "ja": "ja",
+    "en": "en",
+    "zh-CN": "zh-CN",
+    "zh-TW": "zh-TW",
+    "ko": "ko",
+    "fr": "fr",
+    "es": "es",
+    "de": "de",
+    "pt": "pt",
+    "it": "it",
+    "ru": "ru",
+    "nl": "nl",
+    "hi": "hi",
+    "bn": "bn",
+    "vi": "vi",
+    "th": "th",
+    "id": "id",
+    "tr": "tr",
+    "uk": "uk",
+    "pl": "pl",
+    "sw": "sw",
+    "ar": "ar",
+}
+OG_LOCALE_MAP = {
+    "ja": "ja_JP",
+    "en": "en_US",
+    "zh-CN": "zh_CN",
+    "zh-TW": "zh_TW",
+    "ko": "ko_KR",
+    "fr": "fr_FR",
+    "es": "es_ES",
+    "de": "de_DE",
+    "pt": "pt_PT",
+    "it": "it_IT",
+    "ru": "ru_RU",
+    "nl": "nl_NL",
+    "hi": "hi_IN",
+    "bn": "bn_BD",
+    "vi": "vi_VN",
+    "th": "th_TH",
+    "id": "id_ID",
+    "tr": "tr_TR",
+    "uk": "uk_UA",
+    "pl": "pl_PL",
+    "sw": "sw_TZ",
+    "ar": "ar_SA",
+}
+SCHEMA_LANGUAGE_MAP = {
+    "ja": "ja-JP",
+    "en": "en",
+    "zh-CN": "zh-CN",
+    "zh-TW": "zh-TW",
+    "ko": "ko",
+    "fr": "fr",
+    "es": "es",
+    "de": "de",
+    "pt": "pt",
+    "it": "it",
+    "ru": "ru",
+    "nl": "nl",
+    "hi": "hi",
+    "bn": "bn",
+    "vi": "vi",
+    "th": "th",
+    "id": "id",
+    "tr": "tr",
+    "uk": "uk",
+    "pl": "pl",
+    "sw": "sw",
+    "ar": "ar",
+}
+# 言語ごとの代表的な地域 (geo.region は ISO 3166-1 alpha-2)。
+# 元のテンプレートは JP 固定だったが、各言語の主要な利用地域に対応させる。
+GEO_REGION_MAP = {
+    "ja": ("JP", "Japan"),
+    "en": ("US", "United States"),
+    "zh-CN": ("CN", "China"),
+    "zh-TW": ("TW", "Taiwan"),
+    "ko": ("KR", "South Korea"),
+    "fr": ("FR", "France"),
+    "es": ("ES", "Spain"),
+    "de": ("DE", "Germany"),
+    "pt": ("PT", "Portugal"),
+    "it": ("IT", "Italy"),
+    "ru": ("RU", "Russia"),
+    "nl": ("NL", "Netherlands"),
+    "hi": ("IN", "India"),
+    "bn": ("BD", "Bangladesh"),
+    "vi": ("VN", "Vietnam"),
+    "th": ("TH", "Thailand"),
+    "id": ("ID", "Indonesia"),
+    "tr": ("TR", "Türkiye"),
+    "uk": ("UA", "Ukraine"),
+    "pl": ("PL", "Poland"),
+    "sw": ("TZ", "Tanzania"),
+    "ar": ("SA", "Saudi Arabia"),
+}
+LANGUAGE_FALLBACKS = {
+    "ja": (),
+    "en": (),
+    "zh-CN": ("en",),
+    "zh-TW": ("zh-CN", "en"),
+    "ko": ("en",),
+    "fr": ("en",),
+    "es": ("en",),
+    "de": ("en",),
+    "pt": ("en",),
+    "it": ("en",),
+    "ru": ("en",),
+    "nl": ("en",),
+    "hi": ("en",),
+    "bn": ("en",),
+    "vi": ("en",),
+    "th": ("en",),
+    "id": ("en",),
+    "tr": ("en",),
+    "uk": ("en",),
+    "pl": ("en",),
+    "sw": ("en",),
+    "ar": ("en",),
+}
